% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EMEF.R
\name{get_EMEF}
\alias{get_EMEF}
\title{Load EMEF survey data (Desplaçaments or Opinió)}
\usage{
get_EMEF(any_EMEF, tipus = c("desplaçaments", "opinió"))
}
\arguments{
\item{any_EMEF}{Integer or string indicating the survey year
(e.g., \code{2022} or \code{"2022"}).}

\item{tipus}{Character string specifying the dataset type:
either \code{"desplaçaments"} or \code{"opinió"}. Defaults to \code{"desplaçaments"}.}
}
\value{
A data frame containing the loaded EMEF dataset, with an additional
attribute \code{"fitxer_origen"} that records the original filename.
}
\description{
This function loads the raw EMEF survey data (Mobilitat en dia feiner)
from the local directory structure of Institut Metròpoli. It supports
both “desplaçaments” and “opinió” datasets, automatically detecting
the correct \code{.sav} file according to the selected type and year.
}
\details{
The function expects that the data are stored under the following path:
\verb{HOMEPATH/Institut Metròpoli/Mobilitat-BD - General/01_ENQUESTES/EMEF/3-BBDD/}

Depending on the \code{tipus} argument, it will look for one of the following files:
\itemize{
\item \code{"YYYY BBDD DESPLAÇAMENTS EMEF.sav"}
\item \code{"YYYY BBDD OPINIÓ EMEF.sav"}
}

The function reads the SPSS file using \code{foreign::read.spss()}, converts it
into a data frame, and attaches an attribute \code{fitxer_origen} containing the
name of the original file.
}
\examples{
\dontrun{
# Load mobility data
EMEF_despl <- get_EMEF(2022, tipus = "desplaçaments")

# Load opinion data
EMEF_opinio <- get_EMEF(2021, tipus = "opinió")
}

}
